updated text output for possible matches
I've noticed that sometimes it can return a false positive, where it receives a 200 status, but it is not a correct username/pass combo. Added in text to clarify that it is a POSSIBLE match, until I can hunt down actual fix. Timing attack not impacted, just the verification of whether auth was successful.